+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased Changes
 ------------------
+* Issue - Performance optimization: ensure presence and order of instance variables in `PluginOptions` (#3002).
 
 3.191.5 (2024-03-26)
 ------------------